Console.log now works
