+++ test/fallback-errorHandler.js
@@ -6,6 +6,9 @@ const cleanupTestApp = require('./util/cleanupTestApp')
 const path = require('path')
 const appDir = path.join(__dirname, 'app/paramFunctionTest')
 const errorHandlerSandBox = path.join(__dirname, './util/errorHandler.js')
+let errhandler
+let skipDeps = false
+
 const fs = require('fs')
 const { execSync } = require('child_process')
 
@@ -20,7 +23,9 @@ describe('Testing error handlers', function () {
     // Run fallbac k dependancy script
     try {
       execSync(`node ${errorHandlerSandBox}`)
-    } catch (err) { console.log(err) }
+    } catch (err) {
+      console.log(err)
+    }
     this.timeout(20000)
     done()
   })
@@ -36,18 +41,16 @@ describe('Testing error handlers', function () {
     })
   })
 
-  it.only('Checking to see if `fallback-dependencies` is declared in `devDependencies` in your app', function () {
-    const pkg = require('./clones/repo4/package.json')
-    if (pkg.fallbackDependencies.repos === undefined) {
-      // try {
-      //   execSync(`node ${errorHandlerSandBox}`)
-      // } catch (err) { console.log(err) }
-    }
-    const nodeFD = require('./clones/repo4/node_modules/fallback-dependencies')
-
-    console.log(nodeFD)
-    // console.log(pkg.fallbackDependencies.repos)
-
-    assert(pkg.devDependencies !== undefined, '`fallback-dependencies` is not declared in `devDependencies`')
-  })
+  // // Testing line 10 <-- NOT WORKING
+  // it('Checking to see if `fallback-dependencies` still works if repos is non-presence', function () {
+  //   const repo4PKG = require('./clones/repo4/package.json')
+  //   const repo5PKG = require('./clones/repo5/package.json')
+  //   // console.log(process.mainModule)
+  //   assert.deepEqual(repo4PKG.fallbackDependencies.repos, undefined, 'repos is present in package object')
+  //   if (repo4PKG.fallbackDependencies.repos === undefined) {
+  //     assert(errhandler === undefined, 'Repo is non-present in package.json causing an err')
+  //     assert(fs.existsSync(path.join(__dirname, '/clones/repo4/lib')) === false, './test/clones/repo4/lib does exist')
+  //     assert.deepEqual(repo5PKG.fallbackDependencies.repos, {}, 'repos is not an empty object')
+  //   }
+  // })
 })
